requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
backcall==0.2.0
click==8.0.3
colorama==0.4.4
decorator==5.1.0
Flask==2.0.2
Flask-Cors==3.0.10
ipython==7.30.0
itsdangerous==2.0.1
jedi==0.18.1
Jinja2==3.0.3
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
numpy==1.21.4
parso==0.8.2
pickleshare==0.7.5
Pillow==8.4.0
prompt-toolkit==3.0.23
Pygments==2.10.0
six==1.16.0
torch==1.10.0+cpu
torchvision==0.11.1+cpu
traitlets==5.1.1
typing-extensions==4.0.0
wcwidth==0.2.5
Werkzeug==2.0.2
